refactor(admin): profile workspace model 분리
