+++ front/src/pages/admin/profile.tsx
@@ -3,16 +3,13 @@ import { GetServerSideProps, NextPage } from "next"
 import { useRouter } from "next/router"
 import { ChangeEvent, useCallback, useEffect, useMemo, useRef, useState } from "react"
 import { apiFetch, getApiBaseUrl } from "src/apis/backend/client"
-import AppIcon, { IconName } from "src/components/icons/AppIcon"
+import AppIcon from "src/components/icons/AppIcon"
 import ProfileImage from "src/components/ProfileImage"
 import {
-  DEFAULT_CONTACT_ITEM_ICON,
-  DEFAULT_SERVICE_ITEM_ICON,
   getProfileCardIconOptions,
   isAllowedProfileLinkHref,
   normalizeProfileLinkHref,
   ProfileCardLinkItem,
-  ProfileCardLinkSection,
 } from "src/constants/profileCardLinks"
 import { queryKey } from "src/constants/queryKey"
 import useAuthSession, { AuthMember } from "src/hooks/useAuthSession"
@@ -21,7 +18,6 @@ import { setProfileWorkspaceCache, useProfileWorkspace } from "src/hooks/useProf
 import useViewportImageEditor from "src/libs/imageEditor/useViewportImageEditor"
 import {
   buildProfileWorkspaceAdminProfileCacheFields,
-  buildProfileWorkspaceFromLegacy,
   AboutProjectBlock,
   normalizeProfileWorkspaceContent,
   ProfileWorkspaceContent,
@@ -131,15 +127,27 @@ import {
   ModalFooter,
 } from "src/routes/Admin/AdminProfileWorkspace.styles"
 import {
+  WORKSPACE_SECTIONS,
+  buildWorkspaceFallback,
+  createBlankAboutProject,
+  createBlankAboutSection,
+  createBlankLinkItem,
+  type LinkTab,
+  moveListItem,
+  type PreviewMode,
+  reorderListItem,
+  serializeWorkspaceSection,
+  toPayloadLinks,
+  validateLinkInputs,
+  type WorkspaceSectionId,
+} from "src/routes/Admin/AdminProfileWorkspaceModel"
+import {
   AdminWorkspaceActionDockInner,
   AdminWorkspaceHeroCopy,
   AdminWorkspaceHeroLayout,
 } from "src/routes/Admin/AdminSurfacePrimitives"
 
 type NoticeTone = "idle" | "loading" | "success" | "error"
-type WorkspaceSectionId = "identity" | "about" | "home" | "design" | "links"
-type LinkTab = "service" | "contact"
-type PreviewMode = "draft" | "published"
 type OpenIconPicker = `${LinkTab}:${number}` | null
 type ProfileImageDraftTransformState = {
   focusX: number
@@ -161,76 +169,6 @@ type AdminProfileBootstrapPayload = {
 const PROFILE_UNSAVED_CHANGES_MESSAGE = "저장하지 않은 변경 사항이 있습니다. 이 페이지를 떠날까요?"
 const PROFILE_IMAGE_DRAFT_DEFAULT_SOURCE_SIZE: ProfileImageSourceSize = { width: 1, height: 1 }
 const PROFILE_IMAGE_UPLOAD_RETRY_DELAY_MS = 700
-
-const WORKSPACE_SECTIONS: {
-  id: WorkspaceSectionId
-  label: string
-}[] = [
-  {
-    id: "identity",
-    label: "프로필",
-  },
-  {
-    id: "about",
-    label: "About 페이지",
-  },
-  {
-    id: "home",
-    label: "헤더 문구",
-  },
-  {
-    id: "design",
-    label: "디자인",
-  },
-  {
-    id: "links",
-    label: "외부 링크",
-  },
-]
-
-const pickWorkspaceSectionContent = (
-  content: ProfileWorkspaceContent,
-  sectionId: WorkspaceSectionId
-) => {
-  switch (sectionId) {
-    case "identity":
-      return {
-        profileImageUrl: content.profileImageUrl,
-        profileRole: content.profileRole,
-        profileBio: content.profileBio,
-      }
-    case "about":
-      return {
-        aboutHeadline: content.aboutHeadline,
-        aboutRole: content.aboutRole,
-        aboutBio: content.aboutBio,
-        aboutSections: content.aboutSections,
-        aboutProjectSectionTitle: content.aboutProjectSectionTitle,
-        aboutProjects: content.aboutProjects,
-      }
-    case "home":
-      return {
-        blogTitle: content.blogTitle,
-        homeIntroTitle: content.homeIntroTitle,
-        homeIntroDescription: content.homeIntroDescription,
-      }
-    case "design":
-      return {
-        blogDesign: content.blogDesign,
-        legacyBlogScheme: content.legacyBlogScheme,
-      }
-    case "links":
-      return {
-        serviceLinks: content.serviceLinks,
-        contactLinks: content.contactLinks,
-      }
-  }
-}
-
-const serializeWorkspaceSection = (
-  content: ProfileWorkspaceContent,
-  sectionId: WorkspaceSectionId
-) => JSON.stringify(pickWorkspaceSectionContent(normalizeProfileWorkspaceContent(content), sectionId))
 
 const sleep = (ms: number) =>
   new Promise<void>((resolve) => {
@@ -284,126 +222,6 @@ const revalidatePublicBlogAppearance = async (): Promise<boolean> => {
     return response.ok
   } catch {
     return false
-  }
-}
-
-const moveListItem = <T,>(items: T[], index: number, direction: -1 | 1) => {
-  const nextIndex = index + direction
-  if (nextIndex < 0 || nextIndex >= items.length) return items
-  const next = items.slice()
-  const [item] = next.splice(index, 1)
-  next.splice(nextIndex, 0, item)
-  return next
-}
-
-const createLocalId = (prefix: string) => `${prefix}-${Math.random().toString(36).slice(2, 9)}`
-
-const createBlankLinkItem = (section: LinkTab): ProfileCardLinkItem =>
-  section === "service"
-    ? { icon: DEFAULT_SERVICE_ITEM_ICON, label: "", href: "" }
-    : { icon: DEFAULT_CONTACT_ITEM_ICON, label: "", href: "" }
-
-const createBlankAboutSection = (): AboutSectionBlock => ({
-  id: createLocalId("about"),
-  title: "",
-  items: [""],
-  dividerBefore: false,
-})
-
-const createBlankAboutProject = (): AboutProjectBlock => ({
-  id: createLocalId("project"),
-  name: "",
-  summary: "",
-  role: "",
-  href: "",
-  linkLabel: "",
-})
-
-const reorderListItem = <T,>(items: T[], fromIndex: number, toIndex: number) => {
-  if (fromIndex === toIndex) return items
-  if (fromIndex < 0 || toIndex < 0 || fromIndex >= items.length || toIndex >= items.length) {
-    return items
-  }
-
-  const next = items.slice()
-  const [item] = next.splice(fromIndex, 1)
-  next.splice(toIndex, 0, item)
-  return next
-}
-
-const validateLinkInputs = (
-  section: ProfileCardLinkSection,
-  sectionLabel: string,
-  items: ProfileCardLinkItem[]
-): string | null => {
-  for (let index = 0; index < items.length; index += 1) {
-    const item = items[index]
-    const label = item.label.trim()
-    const href = item.href.trim()
-    const rowLabel = `${sectionLabel} ${index + 1}번 링크`
-
-    if (!label && !href) {
-      return `${rowLabel}이 비어 있습니다. 입력하거나 삭제해주세요.`
-    }
-    if (!label || !href) {
-      return `${rowLabel}은 이름과 연결 주소를 모두 입력해야 합니다.`
-    }
-    if (!isAllowedProfileLinkHref(section, href)) {
-      if (section === "service") {
-        return `${rowLabel} 주소는 https:// 또는 http:// 형식만 허용됩니다.`
-      }
-      return `${rowLabel} 주소는 https://, http://, mailto:, tel: 형식만 허용됩니다.`
-    }
-  }
-
-  return null
-}
-
-const toPayloadLinks = (
-  section: ProfileCardLinkSection,
-  items: ProfileCardLinkItem[],
-  defaultIcon: IconName
-): ProfileCardLinkItem[] =>
-  items
-    .map((item) => ({
-      icon: item.icon || defaultIcon,
-      label: item.label.trim(),
-      href: normalizeProfileLinkHref(section, item.href),
-    }))
-    .filter((item) => item.label && item.href)
-
-const buildWorkspaceFallback = (
-  member: AuthMember,
-  initialWorkspace: ProfileWorkspaceResponse | null
-): ProfileWorkspaceResponse => {
-  if (initialWorkspace) {
-    const draft = normalizeProfileWorkspaceContent({
-      ...initialWorkspace.draft,
-      blogDesign: initialWorkspace.draft.blogDesign || member.blogDesign || "legacy",
-      legacyBlogScheme: initialWorkspace.draft.legacyBlogScheme || member.legacyBlogScheme || "dark",
-    })
-    const published = normalizeProfileWorkspaceContent({
-      ...initialWorkspace.published,
-      blogDesign: initialWorkspace.published.blogDesign || member.blogDesign || "legacy",
-      legacyBlogScheme: initialWorkspace.published.legacyBlogScheme || member.legacyBlogScheme || "dark",
-    })
-
-    return {
-      draft,
-      published,
-      lastDraftSavedAt: initialWorkspace.lastDraftSavedAt || member.modifiedAt || null,
-      lastPublishedAt: initialWorkspace.lastPublishedAt || member.modifiedAt || null,
-      dirtyFromPublished: initialWorkspace.dirtyFromPublished,
-    }
-  }
-
-  const content = buildProfileWorkspaceFromLegacy(member)
-  return {
-    draft: content,
-    published: content,
-    lastDraftSavedAt: member.modifiedAt || null,
-    lastPublishedAt: member.modifiedAt || null,
-    dirtyFromPublished: false,
   }
 }
 
@@ -776,8 +594,8 @@ const AdminProfileWorkspacePage: NextPage<AdminProfileWorkspacePageProps> = ({
     () =>
       normalizeProfileWorkspaceContent({
         ...draft,
-        serviceLinks: toPayloadLinks("service", draft.serviceLinks, DEFAULT_SERVICE_ITEM_ICON),
-        contactLinks: toPayloadLinks("contact", draft.contactLinks, DEFAULT_CONTACT_ITEM_ICON),
+        serviceLinks: toPayloadLinks("service", draft.serviceLinks),
+        contactLinks: toPayloadLinks("contact", draft.contactLinks),
       }),
     [draft]
   )
